Feature report v3

+++ moit-v3/back/src/main/java/com/moit/reports/dao/ReportsMapper.java
@@ -25,7 +25,7 @@ public interface ReportsMapper {
 	public List<ReportsDto> selectUserReport( HashMap<String, Object> map );
 
 	// select id="selectUserCnt" resultType="int"
-	public int selectUserCnt(int memberId);
+	public int selectUserCnt(Long memberId);
 	
 	// 사용자 본인이 작성한 신고 내역 상세 조회
 	public ReportsDto selectUserReportDetail(ReportsDto dto);
